moved TextTable to its own file
