+++ Source/swiftlint/Commands/RulesCommand.swift
@@ -10,48 +10,6 @@ import Commandant
 import Result
 import SwiftLintFramework
 
-extension String {
-    private func withPadding(count: Int) -> String {
-        let length = characters.count
-        if length < count {
-            return self +
-                Repeat(count: count - length, repeatedValue: " ").joinWithSeparator("")
-        }
-        return self
-    }
-}
-
-private struct TextTableColumn {
-    let header: String
-    let values: [String]
-
-    var width: Int {
-        return max(header.characters.count, values.reduce(0) { max($0, $1.characters.count) })
-    }
-}
-
-private func fence(strings: [String], separator: String) -> String {
-    return separator + strings.joinWithSeparator(separator) + separator
-}
-
-private struct TextTable {
-    let columns: [TextTableColumn]
-
-    func render() -> String {
-        let separator = fence(columns.map({ column in
-            Repeat(count: column.width + 2, repeatedValue: "-").joinWithSeparator("")
-        }), separator: "+")
-        let header = fence(columns.map({ " \($0.header.withPadding($0.width)) " }),
-            separator: "|")
-        let values = (0..<columns.first!.values.count).map({ rowIndex in
-            fence(columns.map({ column in
-                " \(column.values[rowIndex].withPadding(column.width)) "
-            }), separator: "|")
-        }).joinWithSeparator("\n")
-        return [separator, header, separator, values, separator].joinWithSeparator("\n")
-    }
-}
-
 private let violationMarker = "↓"
 
 struct RulesCommand: CommandType {
@@ -68,25 +26,7 @@ struct RulesCommand: CommandType {
             return .Success()
         }
 
-        let sortedRules = masterRuleList.list.sort { $0.0 < $1.0 }
-        let table = TextTable(columns: [
-            TextTableColumn(header: "identifier", values: sortedRules.map({ $0.0 })),
-            TextTableColumn(header: "opt-in",
-                values: sortedRules.map({ ($0.1.init() is OptInRule) ? "yes" : "no" })),
-            TextTableColumn(header: "correctable",
-                values: sortedRules.map({ ($0.1.init() is CorrectableRule) ? "yes" : "no" })),
-            TextTableColumn(header: "enabled in your config",
-                values: sortedRules.map({
-                    Configuration().rules.map({
-                        $0.dynamicType.description.identifier
-                    }).contains($0.0) ? "yes" : "no"
-                })),
-            TextTableColumn(header: "configuration",
-                values: sortedRules.map({ _, ruleType in
-                    return (ruleType.init() as? _ConfigProviderRule)?.configDescription ?? "N/A"
-                }))
-        ])
-        print(table.render())
+        print(TextTable(ruleList: masterRuleList).render())
         return .Success()
     }
